change to unt128

+++ contracts/interfaces/curve/ICrvEthPool.sol
@@ -17,8 +17,8 @@ interface ICrvEthPool {
     ) external;
 
     function exchange(
-        uint256 i,
-        uint256 j,
+        int128 i,
+        int128 j,
         uint256 dx,
         uint256 min_dy
     ) external payable returns (uint256);
